requirements.txt
langchain
langchain-openai
langchain-pinecone
langchain-community
langchain-experimental
python-dotenv
pinecone
openpyxl